+++ src/main/java/br/com/centroweg/financas/service/dto/investidor/InvestidorRequestDTO.java
@@ -1,9 +1,11 @@
 package br.com.centroweg.financas.service.dto.investidor;
 
+import java.math.BigDecimal;
+
 public record InvestidorRequestDTO(
         String nome,
         String cpf,
-        Double saldoInicial,
+        BigDecimal saldoInicial,
         String tipo
 ) {
 }
