+++ src/main/java/br/com/centroweg/financas/service/dto/investidor/InvestidorResponseDTO.java
@@ -1,10 +1,12 @@
 package br.com.centroweg.financas.service.dto.investidor;
 
+import java.math.BigDecimal;
+
 public record InvestidorResponseDTO(
         Long id,
         String nome,
         String tipo,
-        Double saldo,
+        BigDecimal saldo,
         Integer quantidadeAtivos
 ) {
 }
