+++ src/main/java/br/com/centroweg/financas/service/dto/investimento/InvestimentoResponseDTO.java
@@ -1,11 +1,13 @@
 package br.com.centroweg.financas.service.dto.investimento;
 
+import java.math.BigDecimal;
+
 public record InvestimentoResponseDTO (
         String protocolo,
         String investidorNome,
         String ativoTicker,
-        Double valorTotal,
-        Double impostoRetido,
+        BigDecimal valorTotal,
+        BigDecimal impostoRetido,
         String dataHora
 ){
 }
